Support webkit-test-runner key-value pairs in http tests in DumpRenderTree https://bugs.webkit.org/show_bug.cgi?id=166837
Patch by Antoine Quint <graouts@apple.com> on 2017-01-09
Reviewed by Alex Christensen.

We now correctly handle the --absolutePath argument in DumpRenderTree and use
the provided value when trying to determine the path to the test file. Failing
that, we use the computed URL or the raw path provided to cover all cases.

This was tested in the context of https://bugs.webkit.org/show_bug.cgi?id=165668
by applying this patch and seeing the number test failures dramatically reduced
with WK1 due to the option to turn modern media controls off for older media controls
tests.

* DumpRenderTree/DumpRenderTree.h:
* DumpRenderTree/DumpRenderTreeCommon.cpp:
(parseInputLine):
* DumpRenderTree/TestOptions.h:
* DumpRenderTree/TestOptions.mm:
(TestOptions::TestOptions):
* DumpRenderTree/mac/DumpRenderTree.mm:
(runTest):

Identifier: 183966@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@210517 268f45cc-cd09-0410-ab3c-d52691b4dbfc