'webkit-patch post' no longer works with moved/copied files https://bugs.webkit.org/show_bug.cgi?id=166914 <rdar://problem/29979707>
Reviewed by Daniel Bates.

SVN 1.9 has a bug where when a file has been moved or copied, the svn diff of the new
file outputs only the "Index:" line with an empty body.  This workaround
ignores such output from svn diff.  A issue has been filed with SVN, a link to the
public email list tracking this bug is below.
http://mail-archives.apache.org/mod_mbox/subversion-users/201701.mbox/%3cCF9BDE0A-7454-4405-8259-1120C6B76A03@apple.com%3e

* Scripts/svn-create-patch:
(generateDiff): Ignore an "Index:" line with an empty body.


Identifier: 184174@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@210820 268f45cc-cd09-0410-ab3c-d52691b4dbfc