[GTK] asserting on unknown locale for hyphenation is wrong https://bugs.webkit.org/show_bug.cgi?id=167312
Reviewed by Carlos Garcia Campos.

The fact that we hit the assert on a test called 'hyphenation-unknown-locale' is already
a strong indication we should not have it. In addition to that, Carlos Lopez pointed out
a similar assert was removed from the Mac codepaths when the test was introduced.

* platform/text/hyphen/HyphenationLibHyphen.cpp:
(WebCore::lastHyphenLocation): early return when a locale that is not available is provided
for hyphenation, instead of asserting.

Identifier: 184352@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@211040 268f45cc-cd09-0410-ab3c-d52691b4dbfc