[GTK] ASSERTION FAILED: !m_layerTreeHost in DrawingAreaImpl::display() https://bugs.webkit.org/show_bug.cgi?id=167548
Reviewed by Michael Catanzaro.

The problem is that non accelerated compositing forceRepaint implementation is doing a layout and then calling
display. The layout makes the drawing area enter in AC mode and display asserts that we have a layer tree
host. forceRepaint shouldn't do the layout because display already does that and it correctly handles the case
of entering AC mode during the layout. It shouldn't call setNeedsDisplay either, because that schedules a
display, but we are going to display synchronously.

* WebProcess/WebPage/DrawingAreaImpl.cpp:
(WebKit::DrawingAreaImpl::forceRepaint):

Identifier: 184590@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@211346 268f45cc-cd09-0410-ab3c-d52691b4dbfc