Make Dump Render Tree not crash https://bugs.webkit.org/show_bug.cgi?id=167667
DumpRenderTree is crashing when running WK2 tests that it's not supposed to be running.
DumpRenderTree uses RunLoop::main.dispatch(), but that crashes because it's not set up on iOS.
So we're making use of RunWebThread, which is the correct thing to do on iOS, and it doesn't crash.
This does not fix any tests, it just fixes the crash.

Reviewed by Simon Fraser.

* DumpRenderTree/TestRunner.cpp:
(TestRunner::callUIScriptCallback):



Identifier: 184699@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@211460 268f45cc-cd09-0410-ab3c-d52691b4dbfc